demo/StyleFlow/requirements.txt
dlib==19.21.0
dotmap==1.3.26
imutils==0.5.3
matplotlib==3.1.0
numpy==1.21.4
opencv_python==4.1.0.25
Pillow==9.4.0
PyQt5==5.15.9
qdarkgraystyle==1.0.2
QDarkStyle==2.7
requests==2.24.0
scikit_image==0.15.0
scikit_learn==1.2.1
scipy==1.2.1
skimage==0.0
streamlit==1.2.0
tensorboard==1.15.0
tensorflow==2.11.0
tensorflow_gpu==1.15.0
torch==1.1.0
torchdiffeq==0.0.1
torchvision==0.3.0
tqdm==4.32.1
